Move CONTACTS table to va-telephone story
